Add RandomGen instances and this generalises StdGen
