+++ random-source/random-source.cabal
@@ -51,6 +51,7 @@ Library
   exposed-modules:      Data.Random.Source
                         Data.Random.Source.IO
                         Data.Random.Source.PureMT
+                        Data.Random.Source.RandomGen
                         Data.Random.Source.Std
                         Data.Random.Source.StdGen
                         Data.Random.Internal.Words
